introducao a generics views
