--- .gitgnore
@@ -1,7 +1,0 @@
-# Django #
-*.log
-*.pot
-*.pyc
-__pycache__
-db.sqlite3
-media
