+++ funcionarios/views.py
@@ -1,17 +1,22 @@
-from rest_framework.views import APIView
-from rest_framework.response import Response
-from rest_framework import status
+from rest_framework import generics, mixins
 from .models import Funcionario
 from .serializers import FuncionarioSerializer
 
-class FuncionarioAPIView(APIView):
-    def get(self, request):
-        funcionarios = Funcionario.objects.all()
-        serializer = FuncionarioSerializer(funcionarios, many=True)
-        return Response(serializer.data)
+class FuncionarioAPIView(generics.RetrieveUpdateDestroyAPIView):
+    queryset = Funcionario.objects.all()
+    serializer_class = FuncionarioSerializer
 
-    def post(self, request):
-        serializer = FuncionarioSerializer(data = request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data, status = status.HTTP_201_CREATED)
+class FuncionariosAPIView(generics.ListCreateAPIView):
+    queryset = Funcionario.objects.all()
+    serializer_class = FuncionarioSerializer
+
+class CreateFuncionarioAPIView(mixins.RetrieveModelMixin, mixins.CreateModelMixin, generics.GenericAPIView):
+    queryset = Funcionario.objects.all()
+    serializer_class = FuncionarioSerializer
+
+    def post(self, request, *args, **kwargs):
+        return self.create(request, *args, **kwargs)
+
+
+
+
